requirements.txt
flask
discord-interactions
gunicorn
google-cloud-firestore
